+++ packages/swag-sheet/pages/filters.js
@@ -1,9 +1,78 @@
 import React from 'react'
+import Fakerator from 'fakerator'
 import { css } from '@stitches/react'
-import { StyledContainer, TextField } from '@generates/swag'
-import Uploader from '../components/Uploader.js'
+import { StyledContainer, Button, TextField } from '@generates/swag'
+import useQueryParams from '@generates/use-query-params'
+import Spreadsheet from '../components/Spreadsheet.js'
+
+const initialData = [
+  {
+    'Driver Name': 'Tom Deluge',
+    'Plate Number': 'AX9 920',
+    'Make and Model': 'Subaru Forrester'
+  },
+  {
+    'Driver Name': 'Denise Rich',
+    'Plate Number': 'GR8 LOL',
+    'Make and Model': 'Honda Civic'
+  },
+  {
+    'Driver Name': 'Lenny Jones',
+    'Plate Number': 'TTR 302',
+    'Make and Model': 'Ford F150'
+  }
+]
+
+const faker = Fakerator()
+
+function toFilter (filter) {
+  const [id, value] = filter?.split(':')
+  return { id, value }
+}
+function toFilterString (filter) {
+  return `${filter.id}:${filter.value}`
+}
 
 export default function FiltersPage () {
+  const renderRef = React.useRef(0)
+  const [isLoading, setIsLoading] = React.useState(false)
+  const [data, setData] = React.useState(initialData)
+  const [
+    filters,
+    setFilters
+  ] = useQueryParams('filter', [], f => f?.map(toFilter))
+  const [initialState, setInitialState] = React.useState()
+
+  const onPageIndex = React.useCallback(
+    pageIndex => console.log('Page index', pageIndex),
+    []
+  )
+  const onSortBy = React.useCallback(
+    sortBy => console.log('Sort by', sortBy),
+    []
+  )
+
+  function updateData () {
+    setIsLoading(true)
+    setTimeout(
+      () => {
+        setData(data.map(d => ({ ...d, 'Driver Name': faker.names.name() })))
+        setIsLoading(false)
+      },
+      2000
+    )
+  }
+
+  //
+  React.useEffect(
+    () => {
+      if (renderRef.current === 1) setInitialState({ filters })
+    },
+    [filters]
+  )
+
+  React.useEffect(() => (renderRef.current = renderRef.current + 1), [])
+
   return (
     <StyledContainer className={css({ fontFamily: 'sans-serif' })()}>
 
@@ -11,20 +80,54 @@ export default function FiltersPage () {
         swag-sheet
       </h1>
 
+      <div className={css('div', { marginTop: '2em' })()}>
+        <Button primary onClick={updateData}>
+          Update Data
+        </Button>
+
+        <Button
+          secondary
+          css={{ marginLeft: '1em' }}
+          onClick={() => {
+            setFilters([])
+            setInitialState({ filters: [] })
+          }}
+        >
+          Clear Filters
+        </Button>
+
+      </div>
+
       <div>
-        <Uploader
-          onContinue={data => alert(JSON.stringify(data, undefined, 2))}
-          onCellUpdate={(ctx, value) => alert(value)}
+        <Spreadsheet
           columns={columns => columns.map(col => ({
             ...col,
             disableSortBy: false,
             disableFilters: false,
-            Filter: function Filter () {
-              return <TextField id={col.id} small />
+            Filter: function TextFilter ({ column }) {
+              return <TextField
+                id={col.id}
+                value={column.filterValue || ''}
+                small
+                onChange={evt => {
+                  // Set to undefined to remove the filter entirely.
+                  column.setFilter(evt.target.value || undefined)
+                }}
+              />
             }
           }))}
-          table={{ options: { manualFilters: true } }}
-          onFilter={filters => console.log('Filters', filters)}
+          onPageIndex={onPageIndex}
+          onSortBy={onSortBy}
+          data={data}
+          initialState={initialState}
+          filters={filters}
+          showLoading={true}
+          isLoading={isLoading}
+          onFilter={filters => {
+            if (renderRef.current > 0) {
+              setFilters(filters.map(toFilterString), { update: false })
+            }
+          }}
         />
       </div>