+++ packages/swag/components/fields/SwitchField.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { merge } from '@generates/merger'
 import { motion } from 'framer-motion'
+import { css as stitchesCss } from '@stitches/react'
 import StyledDiv from '../styled/StyledDiv.js'
 import StyledLabel from '../styled/StyledLabel.js'
 
@@ -14,23 +15,24 @@ export default function SwitchField (props) {
   } = props
   const [isOn, setIsOn] = React.useState(value)
 
-  const body = merge(
+  const body = stitchesCss(merge(
     {
+      boxSizing: 'border-box',
       display: 'flex',
       flexShrink: 0,
       cursor: 'pointer',
       alignItems: 'center',
       backgroundColor: '#E2E8F0',
       height: '32px',
-      width: '48px',
+      width: '56px',
       paddingLeft: '4px',
       paddingRight: '4px',
       borderRadius: '32px'
     },
     css.body
-  )
+  ))
 
-  const button = merge(
+  const button = stitchesCss(merge(
     {
       height: '24px',
       width: '24px',
@@ -39,7 +41,7 @@ export default function SwitchField (props) {
       boxShadow: '1px 2px 0 rgba(0,0,0,0.05)'
     },
     css.body
-  )
+  ))
 
   return (
     <StyledDiv
@@ -50,12 +52,12 @@ export default function SwitchField (props) {
 
       <motion.div
         initial={false}
-        style={body}
+        className={body()}
         animate={{ backgroundColor: isOn ? '#22C55E' : '#E2E8F0' }}
       >
         <motion.div
           initial={false}
-          style={button}
+          className={button()}
           animate={{ x: isOn ? 24 : 0 }}
           transition={transition}
         />