+++ packages/swag/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @generates/swag
 
+## 0.0.24
+
+### Patch Changes
+
+- a4b6c1b: Fixing box sizing issue on switchfield
+
+## 0.0.23
+
+### Patch Changes
+
+- 8c93074: Fixing TextField props
+
+## 0.0.22
+
+### Patch Changes
+
+- 81bae45: Removing ultra caches
+
 ## 0.0.21
 
 ### Patch Changes